(refactor, feature): implemented Triangle struct with impl. and fill function

+++ src/matrix.rs
@@ -1,19 +1,19 @@
 pub struct Matrix {
-    rows: [[i16; 3]; 3],
+    rows: [[f32; 3]; 3],
 }
 
 impl Matrix {
-    pub fn new(rows: [[i16; 3]; 3]) -> Self {
+    pub fn new(rows: [[f32; 3]; 3]) -> Self {
         Matrix { rows }
     }
 
-    pub fn get(&self, row: usize, col: usize) -> Option<i16> {
+    pub fn get(&self, row: usize, col: usize) -> Option<f32> {
         let val = self.rows.get(row)?.get(col)?;
 
         Some(*val)
     }
 
-    pub fn det(&self) -> Option<i16> {
+    pub fn det(&self) -> Option<f32> {
         let a = self.get(0, 0)?
             * (self.get(1, 1)? * self.get(2, 2)? - self.get(1, 2)? * self.get(2, 1)?);
